observer: distribute finding tests into their proper test files
Move tests from findings_test.go into the files that test the code
they cover: storage_test.go (H1, M5), metrics_detector_bocpd_test.go
(H2, H3, M6, M7, M8), events_test.go (M1-M4, M9, M10, M12),
stateview_test.go (M11). Helper types moved to test_helpers_test.go.
